requirements.txt
Flask==2.3.3
pandas==2.0.3
scikit-learn==1.3.0
plotly==5.17.0
matplotlib==3.7.2
google-generativeai==0.8.5
python-dotenv==1.1.1
networkx==3.1
numpy==1.24.3